Find .csv files also in subdirectories
Previously we didn't recurse into subdirectories. So the file
~/.lookup/foo.csv would be found but ~/.lookup/foo/bar.csv would not
be. Now both are found.

This is so that .csv files can be maintained in Git
repositories. People can keep a public repo on GitHub and collaborate
on some .csv files. Then clone it onto their own computer and keep it
up to date using git.

Symbolic links are also supported so that you can e.g. keep the repo
in your ~/src directory along with all your other repos and lookup
will still find the .csv files.

Lookup will ignore files, directories and symlinks whose names start
with a dot. This is to prevent recursing into `.git` and `.cache`
directories, and to avoid editor backup files of the `.csv` files.